Creating metric bar component which will be used in detail page also
