+++ ServerlessLibraryAPI/ClientApp/src/components/ItemList/ItemList.scss
@@ -17,9 +17,6 @@
     vertical-align: top;
   }
  
-  // .listcontainer{
-  //   height: 100%;
-  // }
   .libraryitem
   {
     border-bottom: 1px solid #D8D8D8;
@@ -39,17 +36,14 @@
     text-decoration: none; 
   }
 
-  .metrics {
-    font-size: 12px;
-    color: #555555;
-    margin-bottom: 5px;
-  }
-  
   .description {
     font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
     font-size: 14px;
     color: #000000;
     padding-bottom: 10px;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
   }
   strong {
     color: #0072c6;
