+++ ServerlessLibraryAPI/ClientApp/src/components/MetricBar/MetricBar.css
@@ -1,0 +1,5 @@
+.metrics {
+    font-size: 12px;
+    color: #555555;
+    margin-bottom: 5px;
+  }
